Problems

+++ app/assets/javascripts/channels/room.coffee
@@ -11,16 +11,22 @@ App.room = App.cable.subscriptions.create "RoomChannel",
     # Called when there's incoming data on the websocket for this channel
     unless data.content.blank?
       current_user_id = $('#current_user_id').text()
+      user = gon.user
       a = data.user_id
       b = parseInt(current_user_id)
       x = data.image_preview
-      if x is null
-        x = " "
-      $('#messages-table').append '<li id="message-container"><div id="message-infos"><h3>' + data.alias + '</h3> <p> | '+ data.msg_date+'</p></div>
-      <div class="content">' +data.image_link+ '</div> <div class="iframe">' +x+ '</div></li>'
-      if a is b
-        $('#messages-table').append "<div id='delete-btn'><a data-confirm='Are you sure! Do you want delete this post' rel='nofollow' data-method='delete' href='/channels/"+data.channel_id+"/messages/"+data.message_id+"'>delete</a>"
-      scroll_bottom()
+      z = data.current_user_birthday
+
+      console.log(user==z)
+
+      if user == z
+        if x is null
+          x = " "
+        $('#messages-table').append '<li id="message-container"><div id="message-infos"><h3>' + data.alias + '</h3> <p> | '+ data.msg_date+'</p></div>
+        <div class="content">' +data.image_link+ '</div> <div class="iframe">' +x+ '</div></li>'
+        if a is b
+          $('#messages-table').append "<div id='delete-btn'><a data-confirm='Are you sure! Do you want delete this post' rel='nofollow' data-method='delete' href='/channels/"+data.channel_id+"/messages/"+data.message_id+"'>delete</a>"
+        scroll_bottom()
 
 $(document).ready ->
   submit_message()
